Cast before shifting

+++ lib/robust_serial.cpp
@@ -51,19 +51,19 @@ int16_t read_i16(SerialPort &serial_port)
 {
   SerialPort::DataBuffer buffer;
   serial_port.Read(buffer, 2);
-  return (int16_t)(buffer[0] & 0xff) | (int16_t)(buffer[1] << 8 & 0xff00);
+  return (((int16_t) buffer[0]) & 0xff) | (((int16_t) buffer[1]) << 8 & 0xff00);
 }
 
 uint16_t read_u16(SerialPort &serial_port)
 {
   SerialPort::DataBuffer buffer;
   serial_port.Read(buffer, 2);
-  return (uint16_t)(buffer[0] & 0xff) | (uint16_t)(buffer[1] << 8);
+  return (((uint16_t) buffer[0]) & 0xff) | (((uint16_t) buffer[1]) << 8);
 }
 
 int32_t read_i32(SerialPort &serial_port)
 {
   SerialPort::DataBuffer buffer;
   serial_port.Read(buffer, 4);
-  return (int8_t)(buffer[0] & 0xff) | (int8_t)(buffer[1] << 8 & 0xff00) | (int8_t)(buffer[2] << 16 & 0xff0000) | (int8_t)(buffer[3] << 24 & 0xff000000);
+  return (((int32_t) buffer[0]) & 0xff) | (((int32_t) buffer[1]) << 8 & 0xff00) | (((int32_t) buffer[2]) << 16 & 0xff0000) | (((int32_t) buffer[3]) << 24 & 0xff000000);
 }
